# [MRG] Improvements to convert_color_space