… pydicom/data/test_files/SC_ybr_full_uncompressed.dcm