#### What does this implement/fix? Explain your changes.
* Allow use of `YBR_FULL_422` in `convert_color_space` 
* Optimisations for changing colour space
* Minor [docstring update](https://1534-14006067-gh.circle-artifacts.com/0/home/circleci/project/doc/_build/html/reference/generated/pydicom.pixel_data_handlers.util.html#pydicom.pixel_data_handlers.util.convert_color_space)
* Add uncompressed dataset with Photometric Interpretation of YBR_FULL 

#### Benchmarks
Current master
```
TimeConvertColorSpace.time_rgb_ybr           1.79s
TimeConvertColorSpace.time_ybr_rgb           1.63s
```

This PR
```
TimeConvertColorSpace.time_rgb_ybr           1.25s
TimeConvertColorSpace.time_ybr_rgb           1.28s
```