@pep8speaks 2019-08-10
Hello @scaramallion! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:





* In the file [`pydicom/tests/test_handler_util.py`](https://github.com/pydicom/pydicom/blob/836de9e924e26db6b3946dcd97962272c7ea52ca/pydicom/tests/test_handler_util.py):

> [Line 524:9](https://github.com/pydicom/pydicom/blob/836de9e924e26db6b3946dcd97962272c7ea52ca/pydicom/tests/test_handler_util.py#L524): [E124](https://duckduckgo.com/?q=pep8%20E124) closing bracket does not match visual indentation