@scaramallion 2019-08-10
The Travis builds are failing to install Pillow under Python 3.7 for some reason, otherwise everything looks good.

```
Collecting pillow
  Using cached https://files.pythonhosted.org/packages/51/fe/18125dc680720e4c3086dd3f5f95d80057c41ab98326877fc7d3ff6d0ee5/Pillow-6.1.0.tar.gz
Skipping bdist_wheel for pillow, due to binaries being disabled for it.
Installing collected packages: pillow
  Running setup.py install for pillow ... -\|-\|-|/-\-\/\/-\|\-|/-\-|\/\|-/\/done
Successfully installed pillow-6.1.0
Traceback (most recent call last):
  File "<string>", line 1, in <module>
ImportError: cannot import name '_imaging' from 'PIL' (/home/travis/miniconda/envs/testenv/lib/python3.7/site-packages/PIL/__init__.py)
```