@scaramallion 2019-08-15
@darcymason @mrbean-bremen this is ready for review if you haven't seen it