@darcymason 2019-08-16
> @darcymason @mrbean-bremen this is ready for review if you haven't seen it

I'd appreciate someone with more expertise than me to have a look.  Perhaps @rhaxton?