@scaramallion 2019-08-20
For YBR_FULL_422 in PS3.4, C.7.6.3:
>Though the chrominance channels are downsampled, there are still nominally three channels, hence Samples per Pixel (0028,0002) has a value of 3, not 2. I.e., for pixel data in a Native (uncompressed) > format, the Value Length of Pixel Data (7FE0,0010) is not:
>
> Rows (0028,0010) * Columns (0028,0011) * Number of Frames (0028,0008) * Samples per Pixel (0028,0002) * (⌊(Bits Allocated (0028,0100)-1)/8⌋+1)
padded to an even length, as it would otherwise be, but rather is:
>
> Rows (0028,0010) * Columns (0028,0011) * Number of Frames (0028,0008) * 2 * (⌊(Bits Allocated (0028,0100)-1)/8⌋+1) padded to an even length.

Hmm, interesting. Need to fix up the way YBR_FULL_422 bytes are calculated in `get_expected_length`?