@mrbean-bremen 2019-08-20
> Need to fix up the way YBR_FULL_422 bytes are calculated 

Yep, I wondered about this, too (having re-read the standard), but as I can see, you are already on it :)